+++ app.js
@@ -1,38 +1,5 @@
 const http = require("http");
-const fs = require("fs");
-const server = http.createServer((req, res) => {
-  const url = req.url;
-  const method = req.method;
-  if (url === "/") {
-    res.setHeader("Content-Type", "text/html");
-    res.write("<html>");
-    res.write("<head><title>Message</title></head>");
-    res.write(
-      "<body> <form action ='/message' method = 'POST'><input type='text' name='message'><button type='submit'>Send</button></input></form></body>"
-    );
-    res.write("</html>");
-    return res.end();
-  }
-  if (url === "/message" && method === "POST") {
-    const body = [];
-    req.on("data", (chunk) => {
-      console.log(chunk);
-      body.push(chunk);
-    });
-    req.on("end", () => {
-      const parsedBody = Buffer.concat(body).toString();
-      const message = parsedBody.split("=")[1];
-      fs.writeFile("message.txt", message, (err) => {
-        res.statusCode = 302;
-        res.setHeader("Location", "/");
-        return res.end();
-      });
-    });
-  }
-  res.setHeader("Content-Type", "text/html");
-  res.write("<html>");
-  res.write("<head><title>Node-Js</title></head>");
-  res.write("<body> <h1>Hello Form NodeJs Server</h1></body>");
-  res.write("</html>");
-});
+const routes = require("./routes");
+
+const server = http.createServer(routes);
 server.listen(8001);
